refactor: unify loot persistence and commit phases
