+++ src/core/persistence/sqlite/campaign-store.ts
@@ -41,7 +41,7 @@ import { initializeBiomeCatalogSchema } from '../../biomes/biome-catalog.js'
 import { initializeWorldLocationSaveJournalSchema } from '../../worldplanner/world-location-save-journal.js'
 import { initializeGeneratorPresetSchema } from './generator-preset-store.js'
 import { initializeSessionGenerationSchema } from '../../session-generation/generated-run-store.js'
-import { initializeLootSchema } from '../../loot/loot-store.js'
+import { initializeLootSchema } from '../../loot/loot-schema.js'
 import { initializeCharacterLootSchema } from '../../loot/character-loot-store.js'
 import { initializeEncounterPlanSchema } from '../../encounter/encounter-plan-store.js'
 import { initializeSessionPlannerSchema } from '../../session-planner/session-planner-store.js'
